jobs/v4/requirements.txt
google-cloud-talent==0.1.0
